fix: update test expectations and refactor test setup
This change addresses the failing tests in  and  by:
- Adjusting the expected CSV record counts to account for the correct header and data rows.
- Updating the expected output values to match the actual data generated by the  and  commands.
- Refactoring the test setup by moving common interfaces and mock services to  to reduce duplication and improve test maintainability.
